better integrate the MONGODB-X509 authentication tests
