+++ auth.go
@@ -233,7 +233,7 @@ func (socket *mongoSocket) loginClassic(cred Credential) error {
 }
 
 type authX509Cmd struct {
-	Authenticate int
+	Authenticate int // "authenticate" is the mongo command. The value, afaik, is not used.
 	User         string
 	Mechanism    string
 }
