+++ testdb/supervisord-ssl.conf
@@ -1,0 +1,3 @@
+
+[program:ssl1]
+command = mongod --sslMode requireSSL --auth --sslCAFile %(here)s/server.pem --sslPEMKeyFile %(here)s/server.pem --nohttpinterface --noprealloc --nojournal --smallfiles --nssize=1 --oplogSize=1 --dbpath %(here)s/ssl1 --bind_ip=127.0.0.1 --port 40301
